Small fix and move
