+++ src/main/java/hw2/inner/HomeFrame.java
@@ -1,4 +1,4 @@
-package hw2;
+package hw2.inner;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
